% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arashi_cols.R
\name{arashi_cols}
\alias{arashi_cols}
\title{Function to extract Arashi colours as hex codes}
\usage{
arashi_cols(...)
}
\arguments{
\item{...}{Character names of arashi_colours}
}
\description{
This is to create Arashi colour palette for ggplot2. First a function to
extract Arashi colours as hex codes
}
