% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arashi_colour_palettes.R
\name{arashi_pal}
\alias{arashi_pal}
\title{Return function to interpolate an Arashi colour palette}
\usage{
arashi_pal(palette = "arashi", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in arashi_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Return function to interpolate an Arashi colour palette
}
