% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind.stats.R
\name{ind.stats}
\alias{ind.stats}
\title{Compute mean and standard deviation for each group in data with k independent variables. Statistical analysis may be applied for data.}
\usage{
ind.stats(
  i,
  j,
  df,
  latex = FALSE,
  test = c("none", "parametric", "non-parametric"),
  var.equal = NULL,
  digits = 3,
  output = ""
)
}
\arguments{
\item{data}{data}

\item{group}{group}
}
\description{
Compute mean and standard deviation for each group in data with k independent variables. Statistical analysis may be applied for data.
}
