% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jecs_cols.R
\name{jecs_cols}
\alias{jecs_cols}
\title{Function to extract Jecs colours as hex codes}
\usage{
jecs_cols(...)
}
\arguments{
\item{...}{Character names of jecs_colours}
}
\description{
This is to create Jecs colour palette for ggplot2. First a function to
extract Jecs colours as hex codes
}
