% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jecs_colour_palettes.R
\name{jecs_pal}
\alias{jecs_pal}
\title{Return function to interpolate an Jecs colour palette}
\usage{
jecs_pal(palette = "jecs", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in jecs_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Return function to interpolate an Jecs colour palette
}
