% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kruskal.scheffe.R
\name{kruskal.scheffe}
\alias{kruskal.scheffe}
\title{Extended Kruskal-Wallis test followed by Scheffe's test}
\usage{
kruskal.scheffe(data, group = NULL)
}
\arguments{
\item{data}{data}

\item{group}{group}
}
\description{
Extended Kruskal-Wallis test followed by Scheffe's test
}
