% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge2.R
\name{merge2}
\alias{merge2}
\title{Merge multiple dataframes in list with a common key into one dataframe}
\usage{
merge2(dfs, ...)
}
\arguments{
\item{dfs}{dataframe}

\item{...}{other dataframe}
}
\description{
This function can be feed by a list that contains multiple dataframes with a common key. It merges each dataframe in the list using the commong key into a single dataframe.
}
\author{
Shoji F. Nakayama
}
