% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{mjvote}
\alias{mjvote}
\title{Majority voting}
\usage{
mjvote(...)
}
\arguments{
\item{...}{Deprecated. Ignored.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\details{
This function is deprecated and is no longer maintained.
Use the generic MI/brms pipeline instead:
\url{https://github.com/fabregithub/generic-mi-brms-pipeline}.
}
