% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nada.summary.R
\name{nada.summary}
\alias{nada.summary}
\title{Calculating NADA summary statistics}
\usage{
nada.summary(data, cen)
}
\arguments{
\item{data}{data}

\item{cen}{indicator for censor}
}
\description{
Calculating NADA summary statistics
}
