% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub.summary.R
\name{pub.summary}
\alias{pub.summary}
\title{Generating summary statistics for publication}
\usage{
pub.summary(data)
}
\arguments{
\item{data}{data}
}
\description{
Generating summary statistics for publication
}
