% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r4jecs-package.R
\docType{package}
\name{r4jecs-package}
\alias{r4jecs}
\alias{r4jecs-package}
\title{r4jecs: Helper functions for JECS analyses}
\description{
A small collection of helper functions used in JECS analyses, including
summary statistics for left-censored data, numeric formatting helpers, and
ROPE range utilities.
}
\details{
Package metadata and dependencies should be managed through \code{usethis} and
regenerated with \code{roxygen2}, rather than edited manually.
}
\author{
\strong{Maintainer}: Shoji Nakayama \email{fabre@nies.go.jp}

Authors:
\itemize{
  \item Shoji Nakayama \email{fabre@nies.go.jp}
}

Other contributors:
\itemize{
  \item Yukiko Nishihama [contributor]
  \item Chau-Ren Jung [contributor]
}

}
\keyword{internal}
