% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r4jecs.R
\docType{package}
\name{r4jecs}
\alias{r4jecs}
\alias{r4jecs-package}
\title{r4jecs}
\description{
These is a package used in JECS project.
}
\author{
\strong{Maintainer}: Shoji Nakayama \email{fabre@nies.go.jp}

Other contributors:
\itemize{
  \item Yukiko Nishihama [contributor]
  \item Chau-Ren Jung [contributor]
}

}
