% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rndom.R
\name{rndom}
\alias{rndom}
\title{Random forest regression model}
\usage{
rndom(data, val, dp, Formula, N)
}
\arguments{
\item{data}{data}

\item{val}{data for external validation}

\item{dp}{dependent variable}

\item{Formula}{regression formula}

\item{N}{number of trees}
}
\description{
Random forest regression model with 10-fold cross validation and external validation
}
\author{
Jung Chau-Ren
}
