% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rndom2.R
\name{rndom2}
\alias{rndom2}
\title{Random forest regression model}
\usage{
rndom2(data, dp, Formula, N)
}
\arguments{
\item{data}{data}

\item{dp}{dependent variable}

\item{Formula}{regression formula}

\item{N}{number of trees}
}
\description{
Random forest regression model with 10-fold cross validation
}
\author{
Jung Chau-Ren
}
