% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rope_logit.R
\name{rope_llogit}
\alias{rope_llogit}
\title{ROPE range calculation for a logistic model}
\usage{
rope_llogit(prevalence, change = 0.05)
}
\arguments{
\item{prevalence}{prevalence of binomial y variance = 1}

\item{change}{tolerable range}
}
\description{
Calculation of ROPE range for a logistic model.
}
\details{
This function calculates the ROPE range for a logistic model.
}
\examples{
\dontrun{
rope.logit <- rope_llogit(0.3, 0.05)
}

}
\author{
Shoji F. Nakayama
}
