% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpick.R
\name{rpick}
\alias{rpick}
\title{Randomly pick values from fitted distribution}
\usage{
rpick(N, data, model)
}
\arguments{
\item{N}{Number of iteration}

\item{data}{data}

\item{model}{fitted model}
}
\description{
Random picking of values from fitted distribution.
}
\details{
This function picks a value randomly from a fitted distribution.
}
\examples{
\dontrun{
for (i in 1:nrow(df)) {
  df$column[i] <- rpick(N, df$column[i], model)
}}

}
\author{
Shoji F. Nakayama
}
