% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arashi_colour_scale.R
\name{scale_colour_arashi}
\alias{scale_colour_arashi}
\title{Colour scale constructor for Arashi colours}
\usage{
scale_colour_arashi(palette = "arashi", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in arashi_palettes}

\item{discrete}{Boolean indicating whether colour aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale() or
scale_color_gradientn(), used respectively when discrete is TRUE or FALSE}
}
\description{
Colour scale constructor for Arashi colours
}
