% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jecs_colour_scale.R
\name{scale_colour_jecs}
\alias{scale_colour_jecs}
\title{Colour scale constructor for Jecs colours}
\usage{
scale_colour_jecs(palette = "jecs", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in jecs_palettes}

\item{discrete}{Boolean indicating whether colour aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale() or
scale_color_gradientn(), used respectively when discrete is TRUE or FALSE}
}
\description{
Colour scale constructor for Jecs colours
}
