% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arashi_colour_fill.R
\name{scale_fill_arashi}
\alias{scale_fill_arashi}
\title{Fill scale constructor for Arashi colours}
\usage{
scale_fill_arashi(palette = "arashi", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in arashi_palettes}

\item{discrete}{Boolean indicating whether colour aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale() or
scale_fill_gradientn(), used respectively when discrete is TRUE or FALSE}
}
\description{
Fill scale constructor for Arashi colours
}
