% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jecs_colour_fill.R
\name{scale_fill_jecs}
\alias{scale_fill_jecs}
\title{Fill scale constructor for Jecs colours}
\usage{
scale_fill_jecs(palette = "jecs", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in jecs_palettes}

\item{discrete}{Boolean indicating whether colour aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale() or
scale_fill_gradientn(), used respectively when discrete is TRUE or FALSE}
}
\description{
Fill scale constructor for Jecs colours
}
\examples{
\dontrun{
library(ggplot2)
# Fill by discrete variable with different palette + remove legend (guide)
ggplot(mpg, aes(manufacturer, fill = manufacturer)) +
geom_bar() +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
  scale_fill_jecs(palette = 'jecs', guide = 'none')
}

}
