% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summary.R
\name{Summary}
\alias{Summary}
\title{Calculation of summary statistics
Calculating summary statistics}
\usage{
Summary(data)
}
\arguments{
\item{data}{data}
}
\description{
Calculation of summary statistics
Calculating summary statistics
}
\author{
Yukiko Nishihama
}
