% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topology.R
\name{topology}
\alias{topology}
\title{Determination of the number of nodes for SOM}
\usage{
topology(data)
}
\arguments{
\item{data}{data}
}
\description{
Determination of the number of nodes for SOM
}
\author{
Shoji F. Nakayama
}
