% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc.sum.R
\name{desc.sum}
\alias{desc.sum}
\title{Calculation of summary statistics}
\usage{
desc.sum(data)
}
\arguments{
\item{data}{data}
}
\description{
This is a function to generate descriptive statistics of a variable.
}
\author{
Yukiko Nishihama
}
