% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/majority_voting.R
\name{mjvote}
\alias{mjvote}
\title{Majority voting}
\usage{
mjvote(v)
}
\arguments{
\item{v}{data}
}
\description{
The function for providing majority voting function after MICE imputation
}
\details{
This function is used to peform majority voting on data generated by MICE imputation
}
\examples{
\dontrun{
require(r4jecs)
require(tidyverse)
require(miceRanger)
require(doParallel)

set.seed(9973)

# Load data
data(iris)

# Ampute the data. iris contains no missing values by default.
ampIris <- amputeData(iris,perc=0.25)
head(ampIris,10)

# Set up back ends.
cl <- makeCluster(2)
registerDoParallel(cl)

# Perform mice
miceObjPar <- miceRanger(
  ampIris,
  m = 6,
  parallel = TRUE,
  verbose = FALSE
)
stopCluster(cl)
registerDoSEQ()

imp <- completeData(miceObjPar)
imp <- tidyMice(imp)
df.imp <- imp \%>\% group_by(.id) \%>\% summarise_at(.vars = names(.)[3:ncol(imp)], .funs = 'mjvote')

df.imp <- cbind(data.frame(ampIris[, 1]), df.imp)
colnames(df.imp)[1] <- "ID"
df.imp <- df.imp[, -2]
}

}
\author{
Shoji F. Nakayama
}
