% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/majority_voting.R
\name{mjvote}
\alias{mjvote}
\title{Majority voting}
\usage{
mjvote(v)
}
\arguments{
\item{v}{data}
}
\description{
The function for providing majority voting function after MICE imputation
}
\details{
This function is used to peform majority voting on data generated by MICE imputation
}
\examples{
\dontrun{
library(miceRange)
n.core <- detectCores()
cl <- makeCluster(n.core / 2)
registerDoParallel(cl)

df <- readRDS("data.rds")
set.seed(12345)
mice.res <- miceRanger(
  df,
  m = 17,
  maxiter = 10,
  parallel = TRUE,
  returnModels = TRUE,
  verbose = FALSE
  )

stopCluster(cl)
registerDoSEQ()

imp <- completeData(mice.res)

df2 <- tidyMice(imp)
df3 <- df2 \%>\% group_by(.id) \%>\% summarise_at(.vars = names(.)[3:ncols(df2)], .funs = 'mjvote')
df3 <- cbind(data.frame(df[, 1]), df3)
colnames(df3)[1] <- "ID"
df3 <- df3[, -2]
}

}
\author{
Shoji F. Nakayama
}
