% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsf.R
\name{nsf}
\alias{nsf}
\title{Number of significant figures and number below decimal point}
\usage{
nsf(data, NSF, NBD)
}
\arguments{
\item{data}{data}

\item{NSF}{Number of significant figures}

\item{NBD}{Number below decimal point}
}
\description{
Take care of the number of significant figures (NSF) and number below decimal point with the following criteria (NBD).
}
\details{
This sets the number of significant figures as NSF and number below decimal point as NBD. It removes trailing '.' and '0.'
}
\examples{
x <- c(1534, 153.4, 15.34, 1.534, 0.1534, 0.01534, 0.001534, 0.190, 0.200)
nsf(x, 3, 2)
nsf(x, 2, 2)
nsf(x, 3, 3)
nsf(x, 1, 1)

y <- c(5432, 543.2, 54.32, 5.432, 0.5432, 0.05432, 0.005432, NA, NA)
z <- data.frame(X = x, Y = y)
apply(z, 2, nsf, NSF = 3, NBD = 3)

}
\author{
Shoji F. Nakayama
}
