% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsf2.R
\name{nsf2}
\alias{nsf2}
\title{Number of significant figures 2}
\usage{
nsf2(data, digits)
}
\arguments{
\item{data}{data}

\item{digits}{Number of significant figures}
}
\description{
Set the number of significant figures.
}
\details{
This sets the number of significant figures.
}
\examples{
\dontrun{
x <- c(0.00991, 0.00998, 0.28999998, 1.98, 0.01235, 0.0125, 10.2,
         12, 11.2999, 19.9, 99.9, 109, 124, 129, 1999)
nsf2(x, 2)
}

}
\author{
Shoji F. Nakayama
}
