% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub.sum.R
\name{pub.sum}
\alias{pub.sum}
\title{Generate summary statistics for publication}
\usage{
pub.sum(data, useNA = "ifany")
}
\arguments{
\item{data}{data only containing measurement data and censoring information}

\item{useNA}{optional argument to determine whether NA is counted in the table function}
}
\description{
This is a function to generate publication ready summary statistics
}
\author{
Shoji F. Nakayama
}
