% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub.sum.R
\name{pub.sum}
\alias{pub.sum}
\title{Generate summary statistics for publication}
\usage{
pub.sum(data)
}
\arguments{
\item{data}{data}
}
\description{
This is a function to generate publication ready summary statistics
}
\author{
Shoji F. Nakayama
}
