% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub.sum.R
\name{pub.summary}
\alias{pub.summary}
\title{Generate summary statistics for publication
Generating summary statistics for publication}
\usage{
pub.summary(data)
}
\arguments{
\item{data}{data}
}
\description{
Generate summary statistics for publication
Generating summary statistics for publication
}
\author{
Shoji F. Nakayama
}
