% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rope_lm.R, R/roperange.R
\name{rope_lm}
\alias{rope_lm}
\title{ROPE range calculation for a linear model}
\usage{
rope_lm(y)

rope_lm(y)
}
\arguments{
\item{y}{continuous y variable}
}
\description{
Calculation of ROPE range for a linear model.

Calculation of ROPE range for a linear model.
}
\details{
This function calculates the ROPE range for a linear model.

This function calculates the ROPE range for a linear model.
}
\examples{
\dontrun{
rope.lm <- rope_lm(y)
}

\dontrun{
rope.lm <- rope_lm(y)
}

}
\author{
Shoji F. Nakayama
}
