% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_miceRanger.R
\name{tidyMice}
\alias{tidyMice}
\title{Tidy miceRanger results}
\usage{
tidyMice(X)
}
\arguments{
\item{X}{data}
}
\description{
The function for tidy miceRanger imputation results
}
\details{
This function is used to tidy miceRanger imputation results
}
\examples{
\dontrun{
require(tidyverse)
require(miceRanger)
require(doParallel)

set.seed(9973)

# Load data
data(iris)

# Ampute the data. iris contains no missing values by default.
ampIris <- amputeData(iris, perc = 0.25)
head(ampIris, 10)
summary(ampIris)

# Set up back ends.
cl <- makeCluster(2)
registerDoParallel(cl)

# Perform mice
miceObjPar <- miceRanger(
  ampIris,
  m = 6,
  parallel = TRUE,
  verbose = FALSE
)
stopCluster(cl)
registerDoSEQ()

imp <- completeData(miceObjPar)
imp <- tidyMice(imp)
df.imp <- imp \%>\% group_by(.id) \%>\% summarise_at(.vars = names(.)[3:ncol(imp)], .funs = 'mjvote')
df.imp <- df.imp[, -1]

summary(df.imp)
}

}
\author{
Shoji F. Nakayama
}
